requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
adapter-transformers=3.1.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.9.2=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
audioinfo=0.2.0=pypi_0
audioread=3.0.0=pypi_0
backcall=0.2.0=pypi_0
banal=1.0.6=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
commonmark=0.9.1=pypi_0
conda=22.11.1=py38h06a4308_4
conda-package-handling=1.9.0=py38h5eee18b_1
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=10.1.243=h6bb024c_0
curl=7.86.0=h5eee18b_0
dataset=1.6.0=pypi_0
datasets=2.3.2=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.5.1=pypi_0
emoji=2.2.0=pypi_0
executing=1.2.0=pypi_0
expat=2.4.9=h6a678d5_0
filelock=3.8.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.1=pypi_0
fsspec=2022.10.0=pypi_0
gdbm=1.18=hd4cb3f1_4
gdown=4.5.4=pypi_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
git=2.34.1=pl5262hc120c5b_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
glob2=0.7=pypi_0
greenlet=2.0.2=pypi_0
h5py=3.7.0=pypi_0
huggingface-hub=0.7.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
importlib-metadata=5.1.0=pypi_0
importlib-resources=5.10.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.11=pypi_0
ipython=8.7.0=pypi_0
jedi=0.18.2=pypi_0
jiwer=2.5.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9b=h024ee3a_2
jsonnet=0.17.0=py38hfa26641_3
jsonnetbin=0.16.0=pypi_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
levenshtein=0.20.2=pypi_0
libcurl=7.86.0=h91b91d3_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
librosa=0.9.2=pypi_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libxml2=2.9.14=h74e7548_0
llvmlite=0.39.1=pypi_0
lxml=4.9.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
mako=1.2.4=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.7=pypi_0
numba=0.56.4=pypi_0
numpy=1.19.2=pypi_0
numpy-base=1.23.5=py38h31eccc5_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.4=pypi_0
parso=0.8.3=pypi_0
pcre2=10.37=he7ceb23_1
perl=5.34.0=h5eee18b_2
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=py38hace64e9_1
pip=22.3.1=py38h06a4308_0
plotille=4.0.2=pypi_0
pluggy=1.0.0=py38h06a4308_1
pooch=1.6.0=pypi_0
portalocker=2.6.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=3.20.3=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=9.0.0=pypi_0
pycosat=0.6.4=py38h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
pysoundfile=0.9.0.post1=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2022.5=pypi_0
pyyaml=6.0=pypi_0
rapidfuzz=2.13.2=pypi_0
readline=8.2=h5eee18b_0
regex=2022.9.13=pypi_0
requests=2.28.1=py38h06a4308_0
resampy=0.4.2=pypi_0
responses=0.18.0=pypi_0
rich=12.6.0=pypi_0
rouge=1.0.1=pypi_0
rouge-score=0.1.2=pypi_0
ruamel.yaml=0.17.21=py38h5eee18b_0
ruamel.yaml.clib=0.2.6=py38h5eee18b_1
sacrebleu=2.3.1=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.3=pypi_0
sentencepiece=0.1.97=pypi_0
seqeval=1.2.2=pypi_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soundfile=0.11.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.46=pypi_0
sqlite=3.40.0=h5082296_0
stack-data=0.6.2=pypi_0
stanza=1.4.0=pypi_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=py38h06a4308_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu101
tqdm=4.62.3=pypi_0
traitlets=5.7.1=pypi_0
transformers=4.11.3=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
unzip=1.0.0=pypi_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.1.0=pypi_0
xz=5.2.8=h5eee18b_0
yarl=1.8.1=pypi_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zss=1.2.0=pypi_0
zstd=1.4.9=haebb681_0
